fix(guardian): barrel exports, design system, size limits, import order
- Add EmailPanel, NotificationsPanel, PrDiffView to integrations barrel
- Add usePrDiff, useEmail, useNotifications hooks to integrations barrel
- Add CommitHistory and useCommitHistory to git-overview barrel
- Replace raw <p> tags with <Text> from @ui in CommitHistory.tsx
- Extract useCommitHistory into its own file (useGit.ts over 150 lines)
- Split integrations/store.ts into communications-store.ts + github-store.ts hub
- Fix import order in useEmail.ts (zod type import after relative imports)
