+++ src/renderer/features/git-overview/index.ts
@@ -5,12 +5,14 @@
 export { GitPage } from './components/GitPage';
 export { GitStatusCard } from './components/GitStatusCard';
 export { BranchList } from './components/BranchList';
+export { CommitHistory } from './components/CommitHistory';
 export { WorktreeList } from './components/WorktreeList';
 export { CommitPanel } from './components/CommitPanel';
 
 export {
   useGitStatus,
   useGitBranches,
+  useCommitHistory,
   useListWorktrees,
   useCreateBranch,
   useCreateWorktree,
