+++ src/renderer/features/integrations/index.ts
@@ -4,12 +4,15 @@
 
 // Components
 export { IntegrationsPage } from './components/IntegrationsPage';
+export { EmailPanel } from './components/EmailPanel';
 export { GitHubConnectionStatus } from './components/GitHubConnectionStatus';
 export { GitHubPanel } from './components/GitHubPanel';
 export { IssueCreateForm } from './components/IssueCreateForm';
 export { IssueList } from './components/IssueList';
 export { NotificationList } from './components/NotificationList';
+export { NotificationsPanel } from './components/NotificationsPanel';
 export { PrDetailModal } from './components/PrDetailModal';
+export { PrDiffView } from './components/PrDiffView';
 export { PrList } from './components/PrList';
 
 // Store
@@ -29,7 +32,27 @@ export {
   useGitHubPrDetail,
   useGitHubPrs,
   useGitHubRepos,
+  usePrDiff,
 } from './api/useGitHub';
+export {
+  useEmailConfig,
+  useEmailQueue,
+  useRemoveEmailQueued,
+  useRetryEmailQueued,
+  useSendTestEmail,
+  useTestEmailConnection,
+  useUpdateEmailConfig,
+} from './api/useEmail';
+export {
+  useAllNotifications,
+  useMarkAllRead,
+  useMarkNotificationRead,
+  useNotificationsConfig,
+  useStartWatching,
+  useStopWatching,
+  useUpdateNotificationsConfig,
+  useWatcherStatus,
+} from './api/useNotifications';
 
 // Events / Hooks
 export { useIntegrationsEvents } from './hooks/useIntegrationsEvents';
